+++ public/index.html
@@ -38,21 +38,13 @@
 					<li class="nav-item">
 						<a class="nav-link" href="paginas/ayuda.html" style="margin-right: 5px;">Asistencia<i style="margin-left: 5px;"></i></a>
 					</li>
-
-					
-					
-					
-					
 				</ul>
 				<div class="col-6" style="display: flex; justify-content: flex-end;">
 					<input class="form-control" style="margin: 0px;border-radius: 0.375rem 0px 0px 0.375rem;border-right-width: 0px;" type="search" placeholder="Buscar" aria-label="Search">
 					<a class="btn" style="border: 1px solid #dee2e6;border-radius: 0px 0.375rem 0.375rem 0px;background-color: #fff;" href="paginas/busqueda_proveedores.html" type="submit"><i class="icofont-search"></i></a>
 				</div>
 
-				
-
-
-				<li class="nav-item" id="logoutContainer" style="display: none;">
+				<li class="nav-item" id="logoutContainer">
 					<button id="logoutButton" class="btn-logout">Logout</button>
 				</li>
 			</div>
@@ -173,7 +165,7 @@
 
 
 
-    <script type="module" src="logout.js"></script>
+    <script type="module" src="./paginas/authentication/authentication.js"></script>
 
 	<script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
 	<script type="module" src="index.js"></script>
