--- public/login.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" >
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Login - XV Dreams</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="style/style.css"> <!-- Enlace al archivo de estilos -->
-    <script type="module" src="firebaseconfig.js"></script>
-    <script type="module" src="login.js"></script>
-</head>
-<body > <!-- Clase para el fondo del login -->
-    <div class="container">
-        <div class="login-container">
-            <h2> XV Dreams</h2>
-
-            <div id="error-message" class="alert alert-danger" style="display: none;">
-                <!-- Aquí aparecerán los mensajes de error -->
-            </div>
-
-            <form id="login-form">
-                <div class="mb-3">
-                    <label for="email" class="form-label">Email</label>
-                    <input type="email" class="form-control" id="email" required>
-                </div>
-                <div class="mb-3">
-                    <label for="password" class="form-label">Contraseña</label>
-                    <div class="input-group">
-                        <input type="password" class="form-control" id="password" required>
-                        <button type="button" id="toggle-password" class="btn btn-outline-secondary">Mostrar</button>
-                    </div>
-                </div>
-             
-                <button type="submit" class="btn btn-login btn-block">Iniciar sesión</button>
-            </form>
-
-            <p>¿No tienes una cuenta aún? <a href="register.html">Regístrate aquí</a>.</p>
-        </div>
-    </div>
-
-
-
-
-    <script>
-        const loginForm = document.getElementById('login-form');
-        const errorMessageDiv = document.getElementById('error-message');
-
-        loginForm.addEventListener('submit', async (e) => {
-            e.preventDefault();
-            const email = document.getElementById('email').value;
-            const password = document.getElementById('password').value;
-            
-            try {
-                // Call your Firebase login function here
-                // await firebase.auth().signInWithEmailAndPassword(email, password);
-                
-                // Redirect to the homepage or dashboard after successful login
-                // window.location.href = 'homepage.html';
-            } catch (error) {
-                errorMessageDiv.innerText = error.message; // Show error message
-                errorMessageDiv.style.display = 'block'; // Display error message
-            }
-        });
-    </script>
-</body>
-</html>
