+++ public/paginas/authentication/login.html
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html lang="es" >
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>Login - XV Dreams</title>
+	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
+	<link rel="stylesheet" href="style/style.css"> <!-- Enlace al archivo de estilos -->
+	<script type="module" src="firebaseconfig.js"></script>
+	<script type="module" src="authentication.js"></script>
+</head>
+<body >
+	<div class="container">
+		<div class="login-container">
+			<h2> XV Dreams</h2>
+
+			<div id="error-message" class="alert alert-danger" style="display: none;">
+				<!-- Aquí aparecerán los mensajes de error -->
+			</div>
+
+			<form id="login-form">
+				<div class="mb-3">
+					<label for="email" class="form-label">Email</label>
+					<input type="email" class="form-control" id="email" required>
+				</div>
+				<div class="mb-3">
+					<label for="password" class="form-label">Contraseña</label>
+					<div class="input-group">
+						<input type="password" class="form-control" id="password" required>
+						<button type="button" id="toggle-password" class="btn btn-outline-secondary">Mostrar</button>
+					</div>
+				</div>
+			 
+				<button type="submit" class="btn btn-login btn-block">Iniciar sesión</button>
+			</form>
+
+			<p>¿No tienes una cuenta aún? <a href="register.html">Regístrate aquí</a>.</p>
+		</div>
+	</div>
+</body>
+</html>
