+++ public/paginas/authentication/register.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>Register - XV Dreams</title>
+	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
+	<link rel="stylesheet" href="style/style.css">
+	<script type="module" src="firebaseconfig.js"></script>
+	<script type="module" src="authentication.js"></script>
+</head>
+<body class="register-body">
+	<div class="register-container">
+		<h2>XV Dreams</h2>
+
+		<div class="welcome-text">¡Bienvenido! Regístrate para continuar.</div>
+		<div id="error-message" class="alert alert-danger" style="display: none;">
+			<!-- Aquí aparecerán los mensajes de error -->
+		</div>
+		<form id="register-form">
+			<div class="mb-3">
+				<label for="email" class="form-label">Email</label>
+				<input type="email" class="form-control" id="email" required>
+			</div>
+	   
+				
+				<div class="mb-3">
+					<label for="password" class="form-label">Contraseña</label>
+					<div class="input-group">
+						<input type="password" class="form-control" id="password" required>
+						<button type="button" id="toggle-password" class="btn btn-outline-secondary">Mostrar</button>
+						<span class="tooltip-icon" data-tooltip="La contraseña debe tener al menos 6 caracteres, letras y un número.">
+					</div>
+				</div>  
+
+
+			</div>
+			<button type="submit" class="btn btn-primary">Registrarse</button>
+		</form>
+		
+		<p>¿Ya tienes una cuenta? <a href="login.html">Inicia sesión aquí</a>.</p>
+	</div>
+</body>
+</html>
